#20 Has been Added Validation for Delete LeaveType

+++ HR.LeaveManagement.Application/Features/LeaveType/Commands/Delete/DeleteLeaveTypeCommandHandler.cs
@@ -16,6 +16,12 @@ public class DeleteLeaveTypeCommandHandler : IRequestHandler<DeleteLeaveTypeComm
 
     public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
     {
+        DeleteLeaveTypeCommandValidation validation = new();
+        var validationResult=await validation.ValidateAsync(request,cancellationToken);
+
+        if (!validationResult.IsValid)
+            throw new BadRequestException("Invalid Request", validationResult);
+
         var deleteLeaveType = await _repository.GetByIdAsync(request.Id) ?? throw new NotFoundException(nameof(Domain.LeaveType), request.Id);
 
         await _repository.DeleteAsync(deleteLeaveType);
